Prepare the docker container for executing the tests
Configure some path mappings in order that the container can
access the locally downloaded repos and can dump result files
in the local file system.

Also prepare a shell script for runtime of the container,
which does the following for each locally downloaded repo:
Step 1: Copy the repo-archive into the workdir inside the
container
Step 2: Execute the library_tester.py script with the cypthon
installation, that will extract the archive and execute the tests.
Also configure it to later store a summarizing result json file
in the directory /tmp/results/<current date in iso>/<package>,
which gets mapped to the local file system.

TODO: The script execution fails as some custom? imports can't
be handled at the moment (see TODOS in execute_tests.sh).