Fix the GraalPy installation process
Instead of installing tox and virtualenv on the Cypthon base,
we now install it on GraalPy directly and make it globally
available, so Cpython uses it as well.

This does not mean there is anything from GraalPy interfering
with the Cypthon tests. Instead tox and virtualenv just create
respective virtual environments for each tested package,
either with cypthon or graalpy. :)