Prebuild some packages into the venv cache for GraalPy
The basic implementation for that was already given.
I found out that applying the GraalPy patches for the
following packages takes a really long time...
So I decided to just prebuild them while generating
the Docker Image. This way I hope executing the tests
gets a lot faster for packages that rely on one of these
packages :)

- numpy
- matplotlib
- scipy
- scikit-learn
- pandas