Exclude files from content filtering based on pattern
Later, we want to provide the temporary test files for GraalPy for packages,
that ended up with GraalPy test failures for inspection in the browser.
These files are then located inside a "graalpy-tmp" directory,
which should be always shown even if we narrow down the content filter.

For example, if we content-filter with input "AttributeError" to locate packages,
that ended up with errors like this, certainly not all files of the "graalpy-tmp"
directory (so from .tox/graalpylibtest*) include that error, but we do want
these files to appear for inspection of the files, the error originated from.

Also extract the now much more complex filter function to the util file.