Dump tmp ".tox/graalpylibtest*" folder into "results/<package>/graalpy-tmp" folder
If a GraalPy test session was started and ran into failing tests (or with unknown results),
we now also dump the temp test environment files for later inspection in the browser.
This improves the error analysis process, as the files mentioned in potential stacktraces
are now available for detailed inspection.